hide tasks if it related to other section in protection Controller

+++ app/Http/Controllers/sections/ProtectionController.php
@@ -581,8 +581,33 @@ class ProtectionController extends Controller
     {
         $id = $request->invoice_id;
         $tasks = Task::where('id', $id)->first();
-        $tasks->delete();
-        return back();
+
+        //hide tasks if toSecion == 2
+        if ($tasks->toSection == 2) {
+            $tasks->update([
+                'toSection' => null,
+            ]);
+            session()->flash('edit', 'تم   التعديل  بنجاح');
+
+            return back();
+        }
+
+        if ($tasks->fromSection == 2) {
+            $tasks->update([
+                'fromSection' => null,
+            ]);
+            session()->flash('edit', 'تم   التعديل  بنجاح');
+
+            return back();
+        }
+        //if the task is related wont delete but hide it
+        if ($tasks->fromSection == 2 && is_null($tasks->toSection)) {
+            $tasks->delete();
+            session()->flash('edit', 'تم   التعديل  بنجاح');
+
+            return back();
+            return "deleted";
+        }
     }
     public function open_file($id, $file_name)
     {
@@ -886,7 +911,7 @@ class ProtectionController extends Controller
         $tasks = Task::where('id', $id)
             ->where('status', 'completed')
             ->first();
-        $tasks_details = TaskDetails::where('task_id', $id)->where('section_id', 2)->where('status', 'completed')->first();
+        $tasks_details = TaskDetails::where('task_id', $id)->where('status', 'completed')->first();
         $task_attachments = TaskAttachment::where('id_task', $id)->get();
         return view('protection.user.tasks.editReport', compact('tasks', 'tasks_details', 'task_attachments'));
     }
@@ -894,7 +919,7 @@ class ProtectionController extends Controller
     public function submitEditReport($id, Request $request)
     {
         $tasks = Task::where('id', $id)->first();
-        $tasks_details = TaskDetails::where('task_id', $id)->where('section_id', 2)->where('status', 'completed')->first();
+        $tasks_details = TaskDetails::where('task_id', $id)->where('status', 'completed')->first();
 
         $tasks->update([
             'report_status' => 1,
